Dynamic Like button now working

+++ src/main/java/uk/ac/cf/client1/team7sohokidschristmaslights/submissions/ImageRepository_imp.java
@@ -140,7 +140,7 @@ public class ImageRepository_imp implements ImageRepository{
         insert(rating);
         System.out.printf("%n--- Adding review to submission ---%n");
     }
-    public void updateLikeCount(Long id, Integer increment) {
+    public void updateLikeCount(Long id, Short increment) {
         String updateQuery = "UPDATE LikeCounts SET like_count = like_count + ? WHERE submission_id = ?";
         // The increment will be hardcoded to either -1 or 1 during this function's call, depending on the state change of the like button.
         try {
